Added PathTracker class
PathTracker is now used to store recent neural network history. Upon
collision, the recent PathTracker will backpropogate the fact that the
path led to a collision. If no collision happens recently, then it will
backpropogate as a positive.

Need: The car is not turning left for some reason. Appears to be
rb.AngularVelocity is not becoming positive.